Feat/23/service photo (#54)
* #31 fix: modify migrations

* #9 fix : modify selector

* Merge branch '31/fix/server_api' of https://github.com/chaechaeis/UPCY_BE into 31/fix/server_api

* #23 feat : upload photo in s3 socket

---------

+++ services/services.py
@@ -1,7 +1,6 @@
-import io
-import time
 import uuid
-
+import boto3
+import io
 from django.shortcuts import get_list_or_404, get_object_or_404
 from django.db import transaction
 from django.core.files.images import ImageFile
@@ -11,6 +10,8 @@ from django.conf import settings
 from services.models import Service, ServiceKeyword, ServicePhoto, Category, Style, Fit, Texture, Detail
 from .selectors import ServiceSelector
 from users.models import User
+from core.utils import s3_file_upload_by_file_data
+from UpcyProject import settings
 
 
 class ServiceCoordinatorService:
@@ -111,21 +112,36 @@ class ServiceService:
             return service
     
 class ServicePhotoService:
-    def __init__(self):
-        pass
-    
-    @staticmethod
-    def create(image:InMemoryUploadedFile):
-        ext = image.name.split(".")[-1]
-        file_path = '{}.{}'.format(str(time.time())+str(uuid.uuid4().hex),ext)
-        image = ImageFile(io.BytesIO(image.read()),name=file_path)
-        service_photo = ServicePhoto(image=image, service=None)
+    def __init__(self, file):
+        self.file = file
+
+    def upload(self):
+        s3_client = boto3.client(
+            's3',
+            aws_access_key_id = settings.AWS_ACCESS_KEY_ID, 
+            aws_secret_access_key = settings.AWS_SECRET_ACCESS_KEY
+        )
+        extension = self.file.name.split('.')[-1]
+        url = f'services/photo/{uuid.uuid1().hex}.{extension}'
+
+        file_data = io.BytesIO(self.file.read())
+        file_data.seek(0)
+
+        try:
+            s3_client.upload_fileobj(
+                file_data,  # 파일 데이터 객체
+                "cognisle-bucket",
+                url,
+                ExtraArgs={
+                    "ContentType": self.file.content_type
+                }
+            )
+            return f"https://cognisle-bucket.s3.{settings.AWS_S3_REGION_NAME}.amazonaws.com/{url}"
         
-        service_photo.full_clean()
-        service_photo.save()
-        
-        return settings.MEDIA_URL + service_photo.image.name
-    
+        except Exception as e:
+            print(f"Failed to upload file to S3: {e}")
+            return None
+
     @staticmethod
     def process_photos(service:Service, service_photos: list[str]):
         for service_photo in service_photos:
@@ -139,12 +155,57 @@ class ServicePhotoService:
                 print(f"Error in process_photos: {e}")
                 service_photo = None
             
-            if op == 'add':
+            if op == 'add' and service_photo :
                 service_photo.service = service
                 service_photo.full_clean()
                 service_photo.save()
-            elif op == 'remove':
+            elif op == 'remove' and service_photo :
                 service_photo.delete()
+    
+"""
+class ServicePhotoService:
+    def __init__(self):
+        pass
+    
+    @staticmethod
+    def create(image:InMemoryUploadedFile):
+        s3_url = s3_file_upload_by_file_data(
+            upload_file=image,
+            region_name=settings.AWS_S3_REGION_NAME,
+            bucket_name=settings.AWS_STORAGE_BUCKET_NAME,
+            bucket_path='services/photo'  # 고유한 경로 지정
+        )
+
+        if not s3_url:
+            raise Exception("Failed to upload image to S3")
+        
+        service_photo = ServicePhoto(image=s3_url, service=None)
+        service_photo.full_clean()
+        service_photo.save()
+        
+        return service_photo.image #위아래 return 중 뭐가 맞는지 모르겠음
+    
+        return settings.MEDIA_URL + service_photo.image.name #위아래 return 중 뭐가 맞는지 모르겠음
+    @staticmethod
+    def process_photos(service:Service, service_photos: list[str]):
+        for service_photo in service_photos:
+            op, photo_url = service_photo.split(',')
+            
+            photo_path = photo_url.replace(settings.MEDIA_ROOT,'').lstrip('/')
+            
+            try:
+                service_photo, created = ServicePhoto.objects.get_or_create(image=photo_path)
+            except Exception as e:
+                print(f"Error in process_photos: {e}")
+                service_photo = None
+            
+            if op == 'add' and service_photo :
+                service_photo.service = service
+                service_photo.full_clean()
+                service_photo.save()
+            elif op == 'remove' and service_photo :
+                service_photo.delete()
+"""
 
 class ServiceKeywordService:
     def __init__(self):
